separate prompt template
