+++ backend/main.py
@@ -7,20 +7,11 @@ from datetime import datetime
 import os
 
 from service import prompt_to_response
+from prompt_template import build_prompt
 
 load_dotenv()
 
 app = FastAPI()
-
-def build_prompt(data: TravelPlanRequest) -> str:
-    # 150 words only for debugging reason
-    # TODO: This part needs fine tune
-    return (
-        f"You are a travel planning assistant. Please create a {data.budget_level} travel plan "
-        f"for a trip starting on {data.start_date} and ending on {data.end_date}. "
-        f"The user will travel from {data.current_city} to {data.destination_city}. "
-        "Suggest places to visit, daily itinerary, and general travel advice. Keep it under 150 words." 
-    )
 
 @app.post("/plan")
 async def travel_plan_endpoint(request: TravelPlanRequest):
